refactor: replace trait 'Note' with struct NoteSnapshot
This removes the Note trait and replaces it with a struct holding a snapshot in time of a particular note. This also changes some function signatures of the repository trait to better match the intent, e.g. returning Result instead of Option when trying to open a note. It also removes the FileBuffer and instead delegates all responsibility of handling the files to the repository.